bug update
Further testing shows that when the size of `data_vector` > `sampling`, the latter expands to make room for the former. This is expected technical behavior. `sample_means`, which `sampling` affects, is an empty vector. So of course it would expand.

But it's still a problem. The user may want to sample less than the size of `data_vector`.